midi, sysex

+++ src/knob.h
@@ -14,7 +14,7 @@ private:
     uint8_t x, y;
 public:
     void draw(U8G2 &lcd);
-    void setVal(int8_t val) { this->val = val; };
+    void setVal(int16_t val) { this->val = val; };
     Knob(String title, int16_t val, int16_t minVal, int16_t maxVal, uint8_t x, uint8_t y):
         title(title), val(val), minVal(minVal), maxVal(maxVal), x(x), y(y) {};
 };
